# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for المطور platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /social-accounts/telegram/request-otp:
    post:
      operationId: requestTelegramOtp
      tags: [social-accounts]
      summary: Request Telegram OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
  /social-accounts/telegram/verify:
    post:
      operationId: verifyTelegramOtp
      tags: [social-accounts]
      summary: Verify Telegram OTP and connect account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramVerifyInput"
      responses:
        "201":
          description: Account connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/facebook:
    post:
      operationId: addFacebookAccount
      tags: [social-accounts]
      summary: Add Facebook account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacebookAccountInput"
      responses:
        "201":
          description: Account added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/instagram:
    post:
      operationId: addInstagramAccount
      tags: [social-accounts]
      summary: Add Instagram account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramAccountInput"
      responses:
        "201":
          description: Account added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
  /social-accounts/{id}:
    get:
      operationId: getSocialAccount
      tags: [social-accounts]
      summary: Get single social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
    delete:
      operationId: deleteSocialAccount
      tags: [social-accounts]
      summary: Delete social account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /social-accounts/{id}/toggle:
    post:
      operationId: toggleSocialAccount
      tags: [social-accounts]
      summary: Toggle account active/inactive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccount"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    SocialAccount:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        platform:
          type: string
        accountName:
          type: string
        phoneNumber:
          type: ["string", "null"]
        isActive:
          type: boolean
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        followersCount:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, userId, platform, accountName, isActive, status, createdAt]
    TelegramOtpRequest:
      type: object
      properties:
        phoneNumber:
          type: string
      required: [phoneNumber]
    OtpResponse:
      type: object
      properties:
        message:
          type: string
        phoneCodeHash:
          type: string
      required: [message, phoneCodeHash]
    TelegramVerifyInput:
      type: object
      properties:
        phoneNumber:
          type: string
        phoneCode:
          type: string
        phoneCodeHash:
          type: string
      required: [phoneNumber, phoneCode, phoneCodeHash]
    FacebookAccountInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        accountName:
          type: string
      required: [username, password, accountName]
    InstagramAccountInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        accountName:
          type: string
      required: [username, password, accountName]
...